requirements.txt
Flask
Pillow==5.4.1
gevent==1.4.0
gunicorn==19.9.0
keras==2.10.0
tensorflow==2.10.0
numpy
pymongo
